[LEET-00001] Re-organize and add testing
